move the util method to GeneratorUtils
