+++ fusion-endpoint/src/main/java/com/vaadin/fusion/generator/typescript/CodeGeneratorUtils.java
@@ -78,12 +78,4 @@ class CodeGeneratorUtils {
             return node;
         }
     }
-
-    static Object escapeFilePath(Object filePath) {
-        if (filePath != null && filePath instanceof String) {
-            String path = (String) filePath;
-            return path.replaceAll(" ", "%20");
-        }
-        return filePath;
-    }
 }
